requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
chirpstack-api==4.10.2
click==8.1.8
Flask==3.1.0
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
